move validator and factorize unit tests
