+++ geotrek/common/validators.py
@@ -3,11 +3,11 @@ from django.template import Context, Template
 from django.utils.translation import gettext_lazy as _
 
 
-def validate_html_template(template):
+def validate_html_template(value):
     context = Context({"object": {"eid": "123456789"}})
     try:
-        compiled = Template(template)
-        compiled.render(context)  # check Django templating
+        template = Template(value)
+        template.render(context)  # check Django templating
     except Exception as e:
         raise ValidationError(
             _("Incorrect syntax: %(error)s"),
